requirements.txt
python-telegram-bot==20.4
python-dotenv
requests